fix(address-space): break the two import cycles that ESM would reject
Both were fatal under ESM and survivable under CommonJS, where a module in
a cycle sees a partially populated exports object and nothing touches the
missing half until a function runs. Under ESM the bindings are in the
temporal dead zone while the cycle evaluates, and both cycles used one at
module-evaluation time.

namespace_impl builds _constructors_map at module scope from eight
constructors, while ua_method_impl imported _handle_hierarchy_parent back
from it. That function and its _coerce_parent helper move to their own
module, which depends only on base_node_impl and leaf packages.

The eight-file one closed through apply_condition_refresh: both base
classes import it, it imported UAConditionImpl, and UAConditionImpl reaches
back to both through `extends`. It needed that class for a single
instanceof, so the test is now structural. _resend_conditionEvents is
declared only on UAConditionImpl, so it selects exactly what instanceof
selected, and it is the idiom the same function already used one loop below.

Verified with the A&C-02 ConditionRefresh end-to-end test, which drives
that path, plus the address-space suite.
